[frontend] WIP cut toolbar into separate components
